+++ packages/insomnia/vite.config.ts
@@ -2,12 +2,69 @@ import { builtinModules } from 'node:module';
 import path from 'node:path';
 
 import { reactRouter } from '@react-router/dev/vite';
-import { externalDependencies } from 'esbuild.entrypoints';
 import { defineConfig } from 'vite';
 
 import pkg from './package.json';
 import { electronNodeRequire } from './vite-plugin-electron-node-require';
-
+//These will be excluded from the bundle and remain as runtime dependencies
+export const externalDependencies = [
+  '@apideck/better-ajv-errors',
+  '@apidevtools/swagger-parser',
+  '@bufbuild/protobuf',
+  '@connectrpc/connect',
+  '@connectrpc/connect-node',
+  '@getinsomnia/node-libcurl',
+  '@grpc/grpc-js',
+  '@grpc/proto-loader',
+  '@seald-io/nedb',
+  '@segment/analytics-node',
+  '@stoplight/spectral-core',
+  '@stoplight/spectral-formats',
+  '@stoplight/spectral-ruleset-bundler',
+  '@stoplight/spectral-rulesets',
+  'apiconnect-wsdl',
+  'aws4',
+  'chai',
+  'chai-json-schema',
+  'chardet',
+  'clone',
+  'color',
+  'content-disposition',
+  'decompress',
+  'dompurify',
+  'electron-context-menu',
+  'electron-updater',
+  'fastq',
+  'graphql',
+  'graphql-ws',
+  'grpc-reflection-js',
+  'hawk',
+  'hkdf',
+  'hosted-git-info',
+  'html-entities',
+  'http-proxy-agent',
+  'https-proxy-agent',
+  'httpsnippet',
+  'iconv-lite',
+  'isbot',
+  'js-yaml',
+  'jsdom',
+  'jshint',
+  'jsonlint-mod-fixed',
+  'marked',
+  'mime-types',
+  'mocha',
+  'multiparty',
+  'node-forge',
+  'oauth-1.0a',
+  'papaparse',
+  'shell-quote',
+  'socket.io-client',
+  'swagger-ui-dist',
+  'tough-cookie',
+  'uuid',
+  'yaml',
+];
 export default defineConfig(({ mode }) => {
   const __DEV__ = mode !== 'production';
